Now using getProcessorWithoutCaching and moved logics into Config::Impl.
