+++ src/OpenColorIO/CMakeLists.txt
@@ -15,7 +15,6 @@ set(SOURCES
     Caching.cpp
     ColorSpace.cpp
     ColorSpaceSet.cpp
-    ColorSpaceUtils.cpp
     Config.cpp
     Context.cpp
     ContextVariableUtils.cpp
